requirements.txt
Flask
google-generativeai
psycopg2-binary
